+++ bank/models/customer.py
@@ -10,6 +10,6 @@ class Customer(BaseModel):
 
     def __eq__(self, other):
         return self.id == other.id and self.first_name == other.first_name and \
-            self.last_name == other.last_name and self.address = other.address and \
+            self.last_name == other.last_name and self.address == other.address and \
             self.email == other.email
